Convert timestamps to UTC prior to storage.
The previous behaviour was to preserve the assumed local time zone.
This worked out for me with DAViCal and CET winter time zone.
After DST, DAViCal started complaining that "CEST" was not
recognized.  The tzlocal module can, in my case, give a
"valid" key from the Olson database, but there is no guarantee
for that - and I don't want to add tzlocal as a module dependency.
Also, even when using the Olson database, there is no guarantee
that the server will use it.  The RFC says that any usage of
TZID MUST be followed up with a VTIMEZONE component, and I
don't want to bother with that now.